+++ R/ler_dados.R
@@ -164,7 +164,9 @@ datajud_desaninhar_assuntos <- function(dados) {
 #' \dontrun{
 #' cliente <- datajud_cliente()
 #' resposta <- datajud_consultar_processo(
-#'   "0000001-89.2020.8.26.0000", cliente, tribunal = "TJSP"
+#'   "0000001-89.2020.8.26.0000",
+#'   tribunal = "TJSP",
+#'   cliente = cliente
 #' )
 #' datajud_ler_processo(resposta)
 #' }
@@ -211,7 +213,9 @@ datajud_ler_processo <- function(base) {
 #' \dontrun{
 #' cliente <- datajud_cliente()
 #' resposta <- datajud_consultar_processo(
-#'   "0000001-89.2020.8.26.0000", cliente, tribunal = "TJSP"
+#'   "0000001-89.2020.8.26.0000",
+#'   tribunal = "TJSP",
+#'   cliente = cliente
 #' )
 #' datajud_ler_movimentacoes(resposta)
 #' }